+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - [TD-5388] New action to download concepts based on viewing permissions
 - [TD-6959] Add action to allow manage Grant Requests on concept linking view
 
+### Changed
+
+- [TD-7192] Use whitespace tokenizer with word delimiter filter to ignore `_` on text search
+
 ## [7.3.0] 2025-03-18
 
 ### Added